add more tests and removed old ApplicationTest
